feat: check push subscription in account linking modal (#1740)
* refactor: add push subscriptions hook

* feat: check forum subscribed in accoutns modal
