+++ src/hooks/usePushSubscriptions.ts
@@ -1,0 +1,41 @@
+import { useMemo } from 'react'
+
+import { ChainId } from '@dcl/schemas/dist/dapps/chain-id'
+import * as PushAPI from '@pushprotocol/restapi'
+import { useQuery } from '@tanstack/react-query'
+import useAuthContext from 'decentraland-gatsby/dist/context/Auth/useAuthContext'
+
+import { PUSH_CHANNEL_ID } from '../constants'
+import { isSameAddress } from '../entities/Snapshot/utils'
+import { getCaipAddress, getPushNotificationsEnv } from '../utils/notifications'
+
+import { DEFAULT_QUERY_STALE_TIME } from './constants'
+
+export default function usePushSubscriptions() {
+  const [user, userState] = useAuthContext()
+  const chainId = userState.chainId || ChainId.ETHEREUM_GOERLI
+  const env = getPushNotificationsEnv(chainId)
+
+  const { data: subscriptions, isLoading: isLoadingPushSubscriptions } = useQuery({
+    queryKey: [`pushSubscriptions#${user}`],
+    queryFn: () =>
+      user
+        ? PushAPI.user.getSubscriptions({
+            user: getCaipAddress(user, chainId),
+            env,
+          })
+        : null,
+    enabled: !!user,
+    staleTime: DEFAULT_QUERY_STALE_TIME,
+  })
+
+  const isSubscribedToDaoChannel = useMemo(
+    () => !!subscriptions?.find((item: { channel: string }) => isSameAddress(item.channel, PUSH_CHANNEL_ID)),
+    [subscriptions]
+  )
+
+  return {
+    isSubscribedToDaoChannel,
+    isLoadingPushSubscriptions,
+  }
+}
